Remove attempt at transforming on the stream

+++ plenario/models/meta/types.py
@@ -7,6 +7,6 @@ numpy_sqlalchemy_type_map = {
     numpy.datetime64: sqlalchemy.DateTime,
     numpy.float32: sqlalchemy.Float,
     numpy.float64: sqlalchemy.Float,
-    numpy.int64: sqlalchemy.Integer,
+    numpy.int64: sqlalchemy.BigInteger,
     numpy.object_: sqlalchemy.String
 }
